remove `StateTransition`
This should probably have some kind of finite state machine to ensure that the game is always in a valid state, making it impossible to call the wrong function.
As it is now, an invalid state here is an internal logic error that should never happen.  `Result` should generally be used for reporting runtime errors.
Removing it as it's probably more useful to return the `UserError` in some of the functions.

fixup